Separate out land use data access layer
