+++ app/src/api/services/domainLogic/processBuildingUpdate.ts
@@ -2,10 +2,10 @@ import { hasAnyOwnProperty } from '../../../helpers';
 
 import { processCurrentLandUseClassifications } from './currentLandUseClassifications';
 
-export async function processBuildingUpdate(buildingId: number, building: any): Promise<any> {
-    if(hasAnyOwnProperty(building, ['current_landuse_class', 'current_landuse_group', 'current_landuse_order'])) {
-        building = await processCurrentLandUseClassifications(buildingId, building);
+export async function processBuildingUpdate(buildingId: number, buildingUpdate: any): Promise<any> {
+    if(hasAnyOwnProperty(buildingUpdate, ['current_landuse_class', 'current_landuse_group', 'current_landuse_order'])) {
+        buildingUpdate = await processCurrentLandUseClassifications(buildingId, buildingUpdate);
     }
 
-    return building;
+    return buildingUpdate;
 }
